pyproject.toml
[tool.poetry]
name = "breaking_brain_api"
version = "0.1.0"
description = ""
authors = ["Yurii <jurakulek@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
environs = "^9.3.0"
djangorestframework = "^3.12.2"
djangorestframework-simplejwt = "^4.6.0"
djangorestframework-camel-case = "^1.2.0"
flake8 = "^3.8.4"
celery = "^5.0.5"
drf-yasg = "^1.20.0"
mixer = "^7.1.2"
Pillow = "^8.1.0"
mailjet-rest = "^1.3.4"
redis = "^3.5.3"
psycopg2-binary = "^2.8.6"
django-cors-headers = "^3.6.0"
django-filter = "^2.4.0"
argon2 = "^0.1.10"
argon2-cffi = "^20.1.0"
django-nested-inline = "^0.4.2"
django-elasticsearch-dsl = "^7.2.0"
stripe = "^2.58.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
